+++ dictionaries/english-key-dict.mjs
@@ -1,4 +1,4 @@
-export const dictionaryObj = 
+export const englishKeyDict = 
 {
     "A": ".-",
     "B": "-...",