+++ logic/english-to-morse-logic.mjs
@@ -1,19 +1,19 @@
-import { dictionaryObj } from './dictionary.js'
+import { englishKeyDict } from '../dictionaries/english-key-dict.mjs'
 
 
 // ENGLISH TO MORSE
 
-export const searchForMatchingKey = (char) => {
+export const englishToMorseCode = (char) => {
 
     let currentValue = ''
 
     if (char === ' ') {
         currentValue += '/ '
     }
-    const englishKeys = Object.keys(dictionaryObj)
+    const englishKeys = Object.keys(englishKeyDict)
 
     englishKeys.forEach((key) => {
-        const value = dictionaryObj[key]
+        const value = englishKeyDict[key]
 
         if (char === key) {
             currentValue += `${value} `